+++ src/client/src/fluvio.rs
@@ -2,10 +2,10 @@ use std::convert::TryFrom;
 use std::sync::Arc;
 
 use tracing::debug;
-use once_cell::sync::OnceCell;
+use tokio::sync::OnceCell;
+//use once_cell::sync::OnceCell;
 
 use fluvio_socket::{SharedMultiplexerSocket, MultiplexerSocket};
-use fluvio_future::task::run_block_on;
 use fluvio_future::net::DomainConnector;
 //use fluvio_future::native_tls::AllDomainConnector;
 use semver::Version;
@@ -83,12 +83,13 @@ impl Fluvio {
     }
 
     /// lazy get spu pool
-    fn spu_pool(&self) -> Result<Arc<SpuPool>, FluvioError> {
+    async fn spu_pool(&self) -> Result<Arc<SpuPool>, FluvioError> {
         self.spu_pool
-            .get_or_try_init(|| -> Result<Arc<SpuPool>, FluvioError> {
-                let pool = run_block_on(SpuPool::start(self.config.clone(), self.socket.clone()));
+            .get_or_try_init(|| async {
+                let pool = SpuPool::start(self.config.clone(), self.socket.clone()).await;
                 Ok(Arc::new(pool?))
             })
+            .await
             .map(|pool| pool.clone())
     }
 
@@ -114,7 +115,7 @@ impl Fluvio {
     ) -> Result<TopicProducer, FluvioError> {
         let topic = topic.into();
         debug!(topic = &*topic, "Creating producer");
-        Ok(TopicProducer::new(topic, self.spu_pool()?))
+        Ok(TopicProducer::new(topic, self.spu_pool().await?))
     }
 
     /// Creates a new `PartitionConsumer` for the given topic and partition
@@ -145,7 +146,11 @@ impl Fluvio {
     ) -> Result<PartitionConsumer, FluvioError> {
         let topic = topic.into();
         debug!(topic = &*topic, "Creating consumer");
-        Ok(PartitionConsumer::new(topic, partition, self.spu_pool()?))
+        Ok(PartitionConsumer::new(
+            topic,
+            partition,
+            self.spu_pool().await?,
+        ))
     }
 
     /// Provides an interface for managing a Fluvio cluster